requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
cffi==1.17.1
cryptography==44.0.3
cuid2==2.0.1
Django==5.2.1
djangorestframework==3.16.0
dnspython==2.7.0
drf-pydantic==2.7.1
drf-yasg==1.21.10
email_validator==2.2.0
idna==3.10
inflection==0.5.1
packaging==25.0
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
PyMySQL==1.1.1
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
typing-inspection==0.4.0
typing_extensions==4.13.2
uritemplate==4.1.1